+++ src/CarGame.js
@@ -1,26 +1,4 @@
-import readline from "readline";
-
-function readLineAsync(query) {
-  return new Promise((resolve, reject) => {
-    if (arguments.length !== 1) {
-      reject(new Error("arguments must be 1"));
-    }
-
-    if (typeof query !== "string") {
-      reject(new Error("query must be string"));
-    }
-
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-    });
-
-    rl.question(query, (input) => {
-      rl.close();
-      resolve(input);
-    });
-  });
-}
+import { readLineAsync } from "./utils/readLineAsync";
 
 function checkCarName(names) {
   const CAR_NAME_MAX_LENGTH = 5;
